Fix pandas 2.x compatibility issues in Phase 5 data science scripts
Updated 9 scripts to work with pandas 2.x breaking changes:
- Replace deprecated fillna(method='ffill'/'bfill') with ffill()/bfill()
- Remove deprecated fillna(inplace=True), use assignment instead
- Fix TypeError when operating on int64 Series by explicit type casting
- Add group_keys=False to groupby().apply() to prevent MultiIndex
- Fix invalid matplotlib color format ('#lightgray' -> 'lightgray')

All Phase 5 classwork scripts now run successfully with pandas 2.x.

+++ python/learning/05-datascience/classwork/04-data-cleaning/05_scaling_normalization.py
@@ -219,7 +219,7 @@ def main():
 
     # L2 normalization (Euclidean norm)
     print("\nL2 Normalization (unit vector):")
-    df_l2 = df_norm.copy()
+    df_l2 = df_norm.copy().astype(float)
     for idx in df_l2.index:
         row = df_l2.loc[idx]
         l2_norm = np.sqrt((row ** 2).sum())
@@ -236,7 +236,7 @@ def main():
 
     # L1 normalization (Manhattan norm)
     print("\nL1 Normalization:")
-    df_l1 = df_norm.copy()
+    df_l1 = df_norm.copy().astype(float)
     for idx in df_l1.index:
         row = df_l1.loc[idx]
         l1_norm = np.abs(row).sum()
